+++ okitweb/static/okit/pricing/js/resources/object_storage_bucket.js
@@ -20,6 +20,9 @@ class ObjectStorageBucketOciPricing extends OkitOciPricingResource {
     }
 
     getBoM(resource) {
+        const resource_name = resource.getArtifactReference()
+        const bom = {skus: [], price_per_month: 0}
+        return bom
     }
 }